Complete Rust crate split fixtures
