+++ crates/agent-ci-core/Cargo.toml
@@ -12,6 +12,7 @@ workspace = true
 
 [dependencies]
 globset = "0.4"
+serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 serde_yaml = "0.9"
 sha2 = "0.10"
