+++ crates/agent-ci-core/src/lib.rs
@@ -1,4 +1,6 @@
+pub mod events;
 pub mod expr;
 pub mod matrix;
 pub mod plan;
+pub mod state;
 pub mod workflow;
