+++ crates/agent-ci-runtime/src/lib.rs
@@ -3,4 +3,5 @@ pub mod dtu;
 pub mod macos_vm;
 pub mod runner;
 pub mod runner_image;
+pub mod wave;
 pub mod workspace;
