+++ crates/agent-ci/fixtures/docker-socket/README.md
@@ -1,5 +1,3 @@
 # Docker socket fixtures
 
-Future Docker socket resolution contracts shared by TypeScript and Rust. Inputs should capture
-environment variables and probed paths; expected output should include the socket URI and bind
-mount path.
+Docker socket resolution contracts shared by TypeScript and Rust. Inputs capture environment variables and probed paths; expected output includes the socket URI and bind mount path, or expected error substrings.
