+++ crates/agent-ci/fixtures/run-results/README.md
@@ -1,4 +1,3 @@
 # Run result fixtures
 
-Future `run-result.json` contracts. Keep schema version, workflow/job names, statuses, and
-step names; omit volatile durations and absolute log paths.
+Normalized `run-result.json` contracts. Checks keep schema version, workflow/job names, statuses, failing step, and step names while omitting volatile timestamps, durations, SHAs, worktree paths, and log paths.
