+++ crates/agent-ci/src/state.rs
@@ -1,11 +1,14 @@
-use serde::{Deserialize, Serialize};
 use sha1::{Digest, Sha1};
 use std::collections::BTreeMap;
 use std::fs;
 use std::io;
 use std::path::{Path, PathBuf};
 
-pub const RUN_RESULT_SCHEMA_VERSION: u8 = 1;
+pub use agent_ci_core::state::{
+    JobResultInput, JobResultStatus, RUN_RESULT_SCHEMA_VERSION, RunResultFile, RunResultInput,
+    RunResultJobEntry, RunResultStepEntry, StepResultInput, StepResultStatus,
+    build_run_result_json, normalize_run_result,
+};
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct StateDirEnv {
@@ -34,88 +37,6 @@ pub struct LogContext {
     pub log_dir: PathBuf,
     pub output_log_path: PathBuf,
     pub debug_log_path: PathBuf,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RunResultStepEntry {
-    pub name: String,
-    pub status: StepResultStatus,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub log_path: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum StepResultStatus {
-    Passed,
-    Failed,
-    Skipped,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RunResultJobEntry {
-    pub name: String,
-    pub workflow: String,
-    pub status: JobResultStatus,
-    pub duration_ms: u64,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub failing_step: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub debug_log_path: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub steps: Option<Vec<RunResultStepEntry>>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum JobResultStatus {
-    Passed,
-    Failed,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RunResultFile {
-    pub schema_version: u8,
-    pub repo: String,
-    pub branch: String,
-    pub worktree_path: String,
-    pub head_sha: String,
-    pub started_at: String,
-    pub finished_at: String,
-    pub status: JobResultStatus,
-    pub jobs: Vec<RunResultJobEntry>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct RunResultInput {
-    pub repo: String,
-    pub branch: String,
-    pub worktree_path: PathBuf,
-    pub head_sha: String,
-    pub started_at: String,
-    pub finished_at: String,
-    pub results: Vec<JobResultInput>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct JobResultInput {
-    pub name: String,
-    pub workflow: String,
-    pub succeeded: bool,
-    pub duration_ms: u64,
-    pub failing_step: Option<String>,
-    pub debug_log_path: Option<PathBuf>,
-    pub steps: Vec<StepResultInput>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct StepResultInput {
-    pub name: String,
-    pub status: StepResultStatus,
-    pub log_path: Option<PathBuf>,
 }
 
 pub fn resolve_state_dir(env: &StateDirEnv, platform: &str, home_dir: &Path) -> PathBuf {
@@ -288,60 +209,6 @@ pub fn resolve_run_result_path(
     ))
 }
 
-pub fn build_run_result_json(input: &RunResultInput) -> RunResultFile {
-    let jobs = input
-        .results
-        .iter()
-        .map(|result| {
-            let status = if result.succeeded {
-                JobResultStatus::Passed
-            } else {
-                JobResultStatus::Failed
-            };
-            let steps = (!result.steps.is_empty()).then(|| {
-                result
-                    .steps
-                    .iter()
-                    .map(|step| RunResultStepEntry {
-                        name: step.name.clone(),
-                        status: step.status.clone(),
-                        log_path: path_if_exists(step.log_path.as_deref()),
-                    })
-                    .collect()
-            });
-
-            RunResultJobEntry {
-                name: result.name.clone(),
-                workflow: result.workflow.clone(),
-                status,
-                duration_ms: result.duration_ms,
-                failing_step: result.failing_step.clone(),
-                debug_log_path: path_if_exists(result.debug_log_path.as_deref()),
-                steps,
-            }
-        })
-        .collect::<Vec<_>>();
-    let status = if input.results.iter().all(|result| result.succeeded) {
-        JobResultStatus::Passed
-    } else {
-        JobResultStatus::Failed
-    };
-
-    RunResultFile {
-        schema_version: RUN_RESULT_SCHEMA_VERSION,
-        repo: input.repo.clone(),
-        branch: input.branch.clone(),
-        worktree_path: absolute_path(&input.worktree_path)
-            .to_string_lossy()
-            .into_owned(),
-        head_sha: input.head_sha.clone(),
-        started_at: input.started_at.clone(),
-        finished_at: input.finished_at.clone(),
-        status,
-        jobs,
-    }
-}
-
 pub fn write_run_result(input: &RunResultInput, state_dir: Option<&Path>) -> Option<PathBuf> {
     let state_dir = state_dir.map_or_else(default_state_dir_for_process, Path::to_path_buf);
     let file_path =
@@ -378,11 +245,6 @@ fn default_state_dir_for_process() -> PathBuf {
         .map(PathBuf::from)
         .unwrap_or_else(|| PathBuf::from("."));
     resolve_state_dir(&state_env, std::env::consts::OS, &home)
-}
-
-fn path_if_exists(path: Option<&Path>) -> Option<String> {
-    path.filter(|path| path.exists())
-        .map(|path| path.to_string_lossy().into_owned())
 }
 
 fn sanitize_branch(branch: &str) -> String {
